requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
certifi==2018.1.18
chardet==3.0.4
click==6.2
discord.py==0.16.12
frogsay==2.0.0
idna==2.6
multidict==4.1.0
pyasn1==0.1.9
python-RIBBIT==1.0.2.dev0
requests==2.18.4
semidbm==0.5.1
six==1.10.0
urllib3==1.22
websockets==3.4
ws4py==0.4.3
